+++ src/scripts/circleci-job-canceller/Modules/circle_utils.py
@@ -2,16 +2,27 @@ import requests
 import pprint
 
 
+def http_get(url, **args):
+    response = requests.get(url, **args)
+    response.raise_for_status()
+    return response
+
+
+def http_post(url, **args):
+    response = requests.post(url, **args)
+    response.raise_for_status()
+    return response
+
+
 def get_all_items(relative_url, headers, pagination_limit=5):
     url = f"https://circleci.com/api/v2{relative_url}"
     temp_url = str(url)
     pages_iterated = 0
     while True:
-        res = requests.get(temp_url, headers=headers).json()
+        res = http_get(temp_url, headers=headers).json()
         if (res.get("message") == 'An internal server error occurred.'):
             print("Circle side error hit")
             break
-
         try:
             # delegates iteration to the (list), so returns 1 by one...
             yield from res["items"]